langchain_demo/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
chromadb>=0.5.0
ollama>=0.3.0
